+++ vite.config.ts
@@ -15,5 +15,6 @@ export default defineConfig({
     },
     sourcemap: true,
     target: "es2022",
+    emptyOutDir: true,
   },
 });